Push up to resolve Image source errors with TA
