--- components/data.js
@@ -1,43 +1,0 @@
-let profile = [{
-        "profile_pic": "../assets/profile_pic/Nate.jpg",
-        "name": "Nate Ryan",
-        "city": "Portland, OR",
-        "schedule": "06:30p - 09:00p",
-        "typeOfWorkout": "Calisthenics",
-        "music": "The Growlers",
-    },
-    {
-        "profile_pic": "../assets/profile_pic/Blaine.jpg",
-        "name": "Blaine Howitt",
-        "city": "Portland, OR",
-        "schedule": "06:30p - 09:00p",
-        "typeOfWorkout": "Calisthenics",
-        "music": "The Growlers"
-    },
-    {
-        "profile_pic": "../assets/profile_pic/Arlen.png",
-        "name": "Arlen Cornejo",
-        "city": "Portland, OR",
-        "schedule": "06:30p - 09:00p",
-        "typeOfWorkout": "Calisthenics",
-        "music": "The Growlers"
-    },
-    {
-        "profile_pic": "../assets/profile_pic/Herbert.jpg",
-        "name": "Herbert Camposa",
-        "city": "Portland, OR",
-        "schedule": "06:30p - 09:00p",
-        "typeOfWorkout": "Calisthenics",
-        "music": "The Growlers"
-    },
-    {
-        "profile_pic": "../assets/profile_pic/Terence.jpg",
-        "name": "Terence Oddie",
-        "city": "Portland, OR",
-        "schedule": "06:30p - 09:00p",
-        "typeOfWorkout": "Calisthenics",
-        "music": "The Growlers"
-    }
-];
-
-export default profile;
