+++ components/mockDB/data.js
@@ -1,0 +1,48 @@
+let profile = [{
+        "profile_pic": "./mockDB/profile_pic/Nate.jpg",
+        "name": "Nate Ryan",
+        "city": "Portland, OR",
+        "schedule": "06:30p - 09:00p",
+        "typeOfWorkout": "Calisthenics",
+        "music": "The Growlers",
+        "user_note": "Direct trade organic mustache migas, thundercats everyday carry celiac. Marfa wolf sriracha, 90's hashtag vinyl austin kinfolk pug."
+    },
+    {
+        "profile_pic": "./mockDB/profile_pic/Blaine.jpg",
+        "name": "Blaine Howitt",
+        "city": "Portland, OR",
+        "schedule": "06:30p - 09:00p",
+        "typeOfWorkout": "Calisthenics",
+        "music": "The Growlers",
+        "user_note": "Direct trade organic mustache migas, thundercats everyday carry celiac. Marfa wolf sriracha, 90's hashtag vinyl austin kinfolk pug."
+    },
+    {
+        "profile_pic": "./profile_pic/Arlen.png",
+        "name": "Arlen Cornejo",
+        "city": "Portland, OR",
+        "schedule": "06:30p - 09:00p",
+        "typeOfWorkout": "Calisthenics",
+        "music": "The Growlers",
+        "user_note": "Direct trade organic mustache migas, thundercats everyday carry celiac. Marfa wolf sriracha, 90's hashtag vinyl austin kinfolk pug."
+    },
+    {
+        "profile_pic": "./profile_pic/Herbert.jpg",
+        "name": "Herbert Camposa",
+        "city": "Portland, OR",
+        "schedule": "06:30p - 09:00p",
+        "typeOfWorkout": "Calisthenics",
+        "music": "The Growlers",
+        "user_note": "Direct trade organic mustache migas, thundercats everyday carry celiac. Marfa wolf sriracha, 90's hashtag vinyl austin kinfolk pug."
+    },
+    {
+        "profile_pic": "./profile_pic/Terence.jpg",
+        "name": "Terence Oddie",
+        "city": "Portland, OR",
+        "schedule": "06:30p - 09:00p",
+        "typeOfWorkout": "Calisthenics",
+        "music": "The Growlers",
+        "user_note": "Direct trade organic mustache migas, thundercats everyday carry celiac. Marfa wolf sriracha, 90's hashtag vinyl austin kinfolk pug."
+    }
+];
+
+export default profile;
